+++ docs/src/demos/components/index.ts
@@ -5,6 +5,7 @@ export * from './_test_';
 export * from './_test_hook_';
 
 // component previews
+export * from './Accordion';
 export * from './Aside';
 export * from './Aspect';
 export * from './Autocomplete';
@@ -49,3 +50,5 @@ export * from './TextField';
 export * from './Tooltip';
 export * from './VisuallyHidden';
 export * from './XLoader';
+export * from './Tag';
+export * from './DateField';