+++ packages/components/src/index.ts
@@ -1,6 +1,7 @@
 export * from './hooks';
 export * from './theme';
 
+export * from './Accordion';
 export * from './Aside';
 export * from './Aspect';
 export * from './Autocomplete';
@@ -42,8 +43,10 @@ export * from './Table';
 export * from './Text';
 export * from './TextArea';
 export * from './TextField';
+export * from './DateField';
 export * from './Tiles';
 export * from './Tooltip';
+export * from './TagGroup';
 export * from './VisuallyHidden';
 export * from './XLoader';
 export * from './Tabs';