+++ packages/system/src/index.ts
@@ -2,3 +2,4 @@ export * from './components';
 export * from './hooks';
 export * from './keyframes';
 export * from './types';
+export * from './defaultTheme';