+++ themes/theme-core/src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './Accordion.style';
 export * from './Badge.style';
 export * from './Body.style';
 export * from './Button.style';
@@ -6,6 +7,7 @@ export * from './Checkbox.style';
 export * from './Dialog.style';
 export * from './Divider.style';
 export * from './Field.style';
+export * from './DateField.style';
 export * from './Footer.style';
 export * from './Headline.style';
 export * from './HelpText.style';
@@ -23,4 +25,5 @@ export * from './Switch.style';
 export * from './Table.style';
 export * from './Text.style';
 export * from './TextArea.style';
+export * from './Tag.style';
 export * from './Underlay.style';